refactor: move timepicker to lib
